move generateClinicalTasks to scheduling service and use midnight time for clinical timestamp
