+++ src/app/core/services/scheduling.service.ts
@@ -401,4 +401,67 @@ export class SchedulingService {
       console.log(rendered)
     })
   }
+
+  generateClinicalTasks(tasks, associatedTask) {
+    let clinicalTasks = []
+    if (tasks) {
+      clinicalTasks = tasks
+    } else {
+      tasks = []
+    }
+    const protocol = associatedTask.protocol
+    const repeatTimes = this.formatRepeatsAfterClinic(
+      protocol['clinicalProtocol']['repeatAfterClinicVisit']
+    )
+    const now = this.setDateTimeToMidnight(new Date())
+    for (let i = 0; i < repeatTimes.length; i++) {
+      const ts = now.getTime() + repeatTimes[i]
+      const clinicalTask: Task = {
+        index: tasks.length + i,
+        completed: false,
+        reportedCompletion: false,
+        timestamp: ts,
+        name: associatedTask['name'],
+        reminderSettings: protocol['reminders'],
+        nQuestions: associatedTask['questions'].length,
+        estimatedCompletionTime: associatedTask['estimatedCompletionTime'],
+        warning: '',
+        isClinical: true
+      }
+      clinicalTasks.push(clinicalTask)
+    }
+    return this.storage.set(StorageKeys.SCHEDULE_TASKS_CLINICAL, clinicalTasks)
+  }
+
+  formatRepeatsAfterClinic(repeats) {
+    const repeatsInMillis = []
+    const unit = repeats['unit']
+    for (let i = 0; i < repeats['unitsFromZero'].length; i++) {
+      const unitFromZero = repeats['unitsFromZero'][i]
+      switch (unit) {
+        case 'min': {
+          const formatted = unitFromZero * 1000 * 60
+          repeatsInMillis.push(formatted)
+          break
+        }
+        case 'hour': {
+          const formatted = unitFromZero * 1000 * 60 * 60
+          repeatsInMillis.push(formatted)
+          break
+        }
+        case 'day': {
+          const formatted = unitFromZero * 1000 * 60 * 60 * 24
+          repeatsInMillis.push(formatted)
+          break
+        }
+      }
+    }
+    return repeatsInMillis
+  }
+
+  updateTaskToComplete(task): Promise<any> {
+    const updatedTask = task
+    updatedTask.completed = true
+    return this.insertTask(updatedTask)
+  }
 }
